+++ NEWS.md
@@ -1,3 +1,10 @@
+optparse 1.6.5
+==============
+* If the ``callback`` parameter of ``add_option`` / ``make_option`` is not ``NULL`` then
+  the default of ``action`` is now ``"callback"``.  Thanks Greg Minshall for suggestion (#35).
+* Better documentation of ``action=="callback"`` in the man page for ``add_option`` and ``make_option``.
+  Thanks Greg Minshall for bug report (#35).
+
 optparse 1.6.4
 ==============